+++ packages/acme-utils/src/index.tsx
@@ -1,3 +1,3 @@
 export { toSlug } from "./toSlug";
 export { useIsomorphicLayoutEffect } from "./useIsomorphicLayoutEffect";
-export { usePrevious } from "./usePrevious";
+export { usePrevious } from "./use-previous";
